some netCDF files store time not in seconds, but fraction of days?

+++ corelib/crayfish_gdal.h
@@ -93,12 +93,12 @@ protected:
     typedef QHash<QString, QString> metadata_hash; // KEY, VALUE
 
     /* return true on failure */
-    virtual bool parseBandInfo(const metadata_hash& metadata, QString& band_name, int* time) = 0;
+    virtual bool parseBandInfo(const metadata_hash& metadata, QString& band_name, float* time) = 0;
     virtual void determineBandVectorInfo(QString& band_name, bool* is_vector, bool* is_x) = 0;
-    virtual int parseMetadataTime(const QString& time_s);
+    virtual float parseMetadataTime(const QString& time_s);
 
 private:
-    typedef QMap<int, QVector<GDALRasterBandH> > timestep_map; //TIME (sorted), [X, Y]
+    typedef QMap<float, QVector<GDALRasterBandH> > timestep_map; //TIME (sorted), [X, Y]
     typedef QHash<QString, timestep_map > data_hash; //Data Type, TIME (sorted), [X, Y]
 
     void openFile();
